Introduce `dpnp.scipy` submodule (#2603)
This PR proposes introducing `dpnp.scipy` submodule with `linalg` and
`special` providing a SciPy-compatible interface of DPNP and a place for
future SciPy-specific functions.
It moves `dpnp.linalg.lu_factor()`, `dpnp.linalg.lu_solve()` and
`dpnp.special.erf` functions to the new submodule, updates tests and
adds an API Reference section `Routines (SciPy)`
