+++ doc/reference/routines.rst
@@ -1,6 +1,5 @@
---------
-Routines
---------
+Routines (NumPy)
+================
 
 The following pages describe NumPy-compatible routines.
 These functions cover a subset of
