+++ doc/reference/scipy.rst
@@ -1,0 +1,12 @@
+Routines (SciPy) (:mod:`dpnp.scipy`)
+====================================
+
+The following pages describe SciPy-compatible routines.
+These functions cover a subset of
+`SciPy routines <https://docs.scipy.org/doc/scipy/reference/#api-reference>`_.
+
+.. toctree::
+   :maxdepth: 2
+
+   scipy_linalg
+   scipy_special
