+++ doc/reference/scipy_linalg.rst
@@ -1,0 +1,17 @@
+.. currentmodule:: dpnp.scipy.linalg
+
+Linear algebra (:mod:`dpnp.scipy.linalg`)
+=========================================
+
+.. Hint:: `SciPy API Reference: Linear algebra (scipy.linalg) <https://docs.scipy.org/doc/scipy/reference/linalg.html>`_
+
+Decompositions
+--------------
+
+.. autosummary::
+   :toctree: generated/
+   :nosignatures:
+
+   lu
+   lu_factor
+   lu_solve
