+++ dpnp/dpnp_iface_arraycreation.py
@@ -38,6 +38,8 @@ it contains:
 """
 
 
+# pylint: disable=duplicate-code
+
 import operator
 
 import dpctl.tensor as dpt
