+++ dpnp/scipy/special/__init__.py
@@ -25,8 +25,8 @@
 # *****************************************************************************
 
 """
-``dpnp.special``
-================
+``dpnp.scipy.special``
+======================
 
 The submodule provides a large collection of mathematical functions that are
 widely used in science and engineering. It includes special functions of
